Cleanup and fix test coverage
