+++ config/ember-try.js
@@ -1,28 +1,12 @@
+/* jshint node: true */
+
 module.exports = {
   scenarios: [
     {
-      name: 'Ember 1.12',
+      name: 'Ember Canary',
       dependencies: {
-        'ember': '1.12.1'
+        'ember': 'canary'
       }
     },
-    {
-      name: 'Ember 1.13',
-      dependencies: {
-        'ember': '1.13.11'
-      }
-    },
-    {
-      name: 'Ember 2.1',
-      dependencies: {
-        'ember': '2.1.0'
-      }
-    },
-    {
-      name: 'Ember 2.2',
-      dependencies: {
-        'ember': '2.2.0'
-      }
-    }
   ]
 };
