+++ config/environment.js
@@ -1,3 +1,4 @@
+/* jshint node: true */
 'use strict';
 
 module.exports = function(/* environment, appConfig */) {
