+++ tests/acceptance/rendering-components-test.js
@@ -52,7 +52,3 @@ test('using nested components', function(assert) {
     assert.equal(find("#expanded-content").length, 1, "The expanded content is showing");
   });
 });
-
-test('not rendering the application template', function(assert) {
-  assert.equal(find("#application-template").length, 0, "Application template was not rendered");
-});
