+++ tests/dummy/config/environment.js
@@ -16,9 +16,6 @@ module.exports = function(environment) {
     APP: {
       // Here you can pass flags/options to your application instance
       // when it is created
-
-      // This is a flag to supress ember-islands behavior
-      // EMBER_ISLANDS: { bypass: true }
     }
   };
